// src/protos/generated/sbai_tak_protos/sbai_tak_protos/session_request.ts
export enum SessionRequestType {
  SESSION_REQUEST_TYPE_UNSPECIFIED = 0,
  SESSION_REQUEST_TYPE_START = 1,
  SESSION_REQUEST_TYPE_KEEP = 2,
  SESSION_REQUEST_TYPE_TERMINATE = 3,
  UNRECOGNIZED = -1,
}
// ...
export interface SessionRequest {
  requestType: SessionRequestType;
  takId: string;
  robotId: string;
  sequenceNumber?: number | undefined;
  sessionToken?: string | undefined;
}

function createBaseSessionRequest(): SessionRequest {
  return { requestType: 0, takId: "", robotId: "", sequenceNumber: undefined, sessionToken: undefined };
}

export const SessionRequest: MessageFns<SessionRequest> = {
  encode(message: SessionRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.requestType !== 0) {
      writer.uint32(8).int32(message.requestType);
    }
    if (message.takId !== "") {
      writer.uint32(18).string(message.takId);
    }
    if (message.robotId !== "") {
      writer.uint32(26).string(message.robotId);
    }
    if (message.sequenceNumber !== undefined) {
      writer.uint32(32).uint64(message.sequenceNumber);
    }
    if (message.sessionToken !== undefined) {
      writer.uint32(42).string(message.sessionToken);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SessionRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSessionRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.requestType = reader.int32() as any;
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.takId = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.robotId = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.sequenceNumber = longToNumber(reader.uint64());
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.sessionToken = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): SessionRequest {
    return {
      requestType: isSet(object.requestType) ? sessionRequestTypeFromJSON(object.requestType) : 0,
      takId: isSet(object.takId) ? globalThis.String(object.takId) : "",
      robotId: isSet(object.robotId) ? globalThis.String(object.robotId) : "",
      sequenceNumber: isSet(object.sequenceNumber) ? globalThis.Number(object.sequenceNumber) : undefined,
      sessionToken: isSet(object.sessionToken) ? globalThis.String(object.sessionToken) : undefined,
    };
  },

  toJSON(message: SessionRequest): unknown {
    const obj: any = {};
    if (message.requestType !== 0) {
      obj.requestType = sessionRequestTypeToJSON(message.requestType);
    }
    if (message.takId !== "") {
      obj.takId = message.takId;
    }
    if (message.robotId !== "") {
      obj.robotId = message.robotId;
    }
    if (message.sequenceNumber !== undefined) {
      obj.sequenceNumber = Math.round(message.sequenceNumber);
    }
    if (message.sessionToken !== undefined) {
      obj.sessionToken = message.sessionToken;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<SessionRequest>, I>>(base?: I): SessionRequest {
    return SessionRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<SessionRequest>, I>>(object: I): SessionRequest {
    const message = createBaseSessionRequest();
    message.requestType = object.requestType ?? 0;
    message.takId = object.takId ?? "";
    message.robotId = object.robotId ?? "";
    message.sequenceNumber = object.sequenceNumber ?? undefined;
    message.sessionToken = object.sessionToken ?? undefined;
    return message;
  },
};

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}